Handle admin scenario config on client
